fix failing unit tests

+++ applicationset/services/pull_request/utils_test.go
@@ -418,8 +418,8 @@ func TestCreatedWithinFilter(t *testing.T) {
 	pullRequests, err := ListPullRequests(t.Context(), provider, filters)
 	require.NoError(t, err)
 	assert.Len(t, pullRequests, 2)
-	assert.Equal(t, 1, pullRequests[0].Number)
-	assert.Equal(t, 3, pullRequests[1].Number)
+	assert.Equal(t, int64(1), pullRequests[0].Number)
+	assert.Equal(t, int64(3), pullRequests[1].Number)
 }
 
 func TestUpdatedWithinFilter(t *testing.T) {
@@ -467,7 +467,7 @@ func TestUpdatedWithinFilter(t *testing.T) {
 	pullRequests, err := ListPullRequests(t.Context(), provider, filters)
 	require.NoError(t, err)
 	assert.Len(t, pullRequests, 1)
-	assert.Equal(t, 3, pullRequests[0].Number)
+	assert.Equal(t, int64(3), pullRequests[0].Number)
 }
 
 func TestCreatedUpdatedWithinComboFilter(t *testing.T) {
@@ -536,8 +536,8 @@ func TestCreatedUpdatedWithinComboFilter(t *testing.T) {
 	pullRequests, err := ListPullRequests(t.Context(), provider, filters)
 	require.NoError(t, err)
 	assert.Len(t, pullRequests, 2)
-	assert.Equal(t, 1, pullRequests[0].Number)
-	assert.Equal(t, 4, pullRequests[1].Number)
+	assert.Equal(t, int64(1), pullRequests[0].Number)
+	assert.Equal(t, int64(4), pullRequests[1].Number)
 }
 
 func TestNoFilters(t *testing.T) {
